Fixing issue with archives returning paywalled content
Originally I would get the first link that archive.ph provides, but for some articles from Atlantic that would lead to the content being cut off and paywalled. Instead of the first link in the first row, archive.ph has a text block on the right with the recommended link and I use that instead.